refactor: use bun + vite
