+++ package.json
@@ -1,14 +1,13 @@
 {
-  "name": "fpl.lol",
-  "version": "0.1.0",
+  "name": "fpl.zoid.dev",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "test": "tsc --noEmit",
-    "code:prettify": "prettier . --write"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "test": "tsc --noEmit"
   },
   "dependencies": {
     "@geist-ui/react": "2.2.5",
@@ -17,6 +16,7 @@
     "next": "13.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-helmet": "^6.1.0",
     "react-no-ssr": "1.1.0",
     "sql-formatter": "4.0.2",
     "ts-pattern": "4.2.1"
@@ -25,9 +25,12 @@
     "@trivago/prettier-plugin-sort-imports": "4.1.1",
     "@types/base-64": "1.0.0",
     "@types/node": "18.14.6",
-    "@types/react": "18.0.28",
+    "@types/react": "18.0.27",
+    "@types/react-dom": "18.0.10",
+    "@types/react-helmet": "^6.1.6",
     "@types/react-no-ssr": "1.1.3",
     "@types/sql.js": "1.4.4",
+    "@vitejs/plugin-react": "3.1.0",
     "arg": "5.0.2",
     "autoprefixer": "10.4.13",
     "eslint": "8.35.0",
@@ -36,7 +39,8 @@
     "prettier": "2.8.4",
     "tailwindcss": "3.2.7",
     "tasuku": "2.0.1",
-    "typescript": "4.9.5",
+    "typescript": "4.9.3",
+    "vite": "4.1.0",
     "zx": "7.2.0"
   }
 }
