+++ vite.config.ts
@@ -1,0 +1,8 @@
+import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react()],
+  assetsInclude: ['**/*.db'],
+})
